xserver Add PRIME flipping functions to ScreenRec
Adds typdefs for (*StartFlippingPixmapTrackingProcPtr),
(*PresentTrackedFlippingPixmapProcPtr), (*StopFlippingPixmapTrackingProcPtr),
(*InitSharedPixmapFlippingProcPtr), (and (*FiniSharedPixmapFlippingProcPtr).

Adds (Init/Fini)SharedPixmapFlipping, (Start/Stop)FlippingPixmapTracking, and
PresentTrackedFlippingPixmap to ScreenRec.

InitSharedPixmapFlipping will allow the sink driver to setup fencing for
flipping between two shared pixmaps, along with the page flip event handler to
ask the source driver for another present in response to a flip.

FiniSharedPixmapFlipping will allow the sink driver to do teardown associated
with InitSharedPixmapFlipping. This will likely involve aborting page flip event
handlers, but the source driver will be responsible for tearing down a buffer's
reservation object, along with any outstanding fences.

(Start/Stop)FlippingPixmapTracking are merely the double-buffered equivalents of
(Start/Stop)PixmapTracking, allowing the source driver to do whatever setup and
teardown necessary for presenting on the two shared pixmaps.

PresentTrackedFlippingPixmap is a function exposed by the source driver for the
X server or sink driver to call to request a present on a given shared pixmap.
This way, presents can be driven by the sink's vblank instead of a timer or
similar mechanism.