xserver Factor out shared pixmap creation
The old version of rrCreateSharedPixmap(), in addition to actually creating a
shared pixmap with scanout, also set up pixmap tracking on the source driver.

I will be needing to create multiple shared pixmaps for PRIME double buffering,
so factor the part that does shared pixmap creation into its own function, the
new rrCreateSharedPixmap(). Rename the old rrCreateSharedPixmap() to
rrSetupPixmapSharing(), a function that replicates the old functionality of
rrCreateSharedPixmap() using the new rrCreateSharedPixmap().