xserver Factor out shared pixmap destruction
Shared pixmap destruction is done by unrefing the master pixmap twice: once for
the original reference, and once for the reference implicitly added by
PixmapShareToSlave. Then, unrefing the slave pixmap once.

When I add PRIME double buffering and synchronization, I will need to do this in
multiple places. To avoid duplication of code and comments explaining it
everywhere, factor it out into its own function and use that in place of where
it was before.