modesetting Fix mmap leak in scanout_pixmap_cpu
drmmode_set_scanout_pixmap_cpu() uses drmmode_map_slave_bo(), which maps a dumb
bo. However, when cleaning up scanout pixmaps, it never destroys it, leading to
a leak of mapped memory. With enough modesets, this can result in an insane
amount of memory being mapped. Add dumb_bo_destroy() of the backing bo as part
of the cleanup process.